# docs/swagger.yaml
basePath: /
definitions:
  handlers.changePasswordRequest:
    properties:
      password:
        type: string
    type: object
  handlers.createRolesRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phoneNumber:
        type: integer
    type: object
...
  handlers.rolesResponse:
    properties:
      birthday:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phonenumber:
        type: integer
    type: object
  handlers.updateRolesRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      name:
        type: string
      phoneNumber:
        type: integer
    type: object
  handlers.updateUserRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      name:
        type: string
      phoneNumber:
        type: integer
    type: object
  handlers.userResponse:
    properties:
      birthday:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phonenumber:
        type: integer
    type: object
...
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: "OZINSHE\tAPI"
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.rolesResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get roles list
      tags:
      - roles - это выполняет роль Админа
    post:
      consumes:
      - application/json
      parameters:
      - description: Roles data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create role
      tags:
      - roles - это выполняет роль Админа
  /roles/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete role
      tags:
      - roles - это выполняет роль Админа
    get:
      consumes:
      - application/json
      parameters:
      - description: Roles id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.rolesResponse'
            type: array
        "400":
          description: Invalid role id
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find role by id
      tags:
      - roles - это выполняет роль Админа
    put:
      consumes:
      - application/json
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: integer
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.updateRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update role
      tags:
      - roles - это выполняет роль Админа
  /roles/{id}/changePassword:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: integer
      - description: Password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Change role password
      tags:
      - roles - это выполняет роль Админа
  /rolesmovie:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of movies
          schema:
            items:
              $ref: '#/definitions/models.MovieAdminResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all moviesAdmin
      tags:
      - Админ получает список фильмов
  /rolesmovie/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete movie
      tags:
      - Админ удаляет фильм
    get:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieAdminResponse'
        "400":
          description: Invalid Movie Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find by id
      tags:
      - Админ ищет фильм
    put:
      consumes:
      - application/json
      parameters:
      - description: Title of the movie
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Description of the movie
        in: body
        name: description
        required: true
        schema:
          type: string
      - description: ReleaseYear of the movie
        in: body
        name: releaseYear
        required: true
        schema:
          type: integer
      - description: Director
        in: body
        name: director
        required: true
        schema:
          type: string
      - description: TrailerUrl
        in: body
        name: trailerUrl
        required: true
        schema:
          type: string
      - description: Genre ids
        in: body
        name: genreIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Category ids
        in: body
        name: categoryIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Age ids
        in: body
        name: ageIds
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Could not bind json
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update moviesAdmin
      tags:
      - Админ редактирует фильм
  /rolesmovie/{id}/setWatched:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      - description: Flag value
        in: query
        name: isWatched
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Mark moviesAdmin as watched
      tags:
      - количество просмотров доступен только для Админа
  /rolesuser:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.userResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Roles gets user's list
      tags:
      - Админ получает список юзеров
  /rolesuser/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete user
      tags:
      - Админ удаляет юзера
    put:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Role Updates user´s information
      tags:
      - Админ редактирует данные юзера
  /rolesuser/{id}/changePassword:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: Password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Change user password
      tags:
      - Админ меняет пароль юзера
  /rolesusers/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Roles id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.rolesResponse'
            type: array
        "400":
          description: Invalid role id
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Role Finds user by id
      tags:
      - поиск user по id(Админ ищет юзера)
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"